fixes issue with integer and biginteger, adds return json string as file uploaded.

+++ db/migrate/20181118103657_add_fields_to_items.rb
@@ -4,7 +4,7 @@ class AddFieldsToItems < ActiveRecord::Migration[5.2]
     add_column :items, :brand, :string
     add_column :items, :model, :string
     add_column :items, :operable, :boolean
-    add_column :items, :checkout_user_id, :integer
+    add_column :items, :checkout_user_id, :bigint
     add_foreign_key :items, :users, column: :checkout_user_id, primary_key: :id
   end
 end
